requirements.txt
aiogram==3.13.1
aiosqlite==0.20.0
APScheduler==3.10.4
openai==1.54.3
httpx==0.27.2
beautifulsoup4==4.12.3
python-dotenv==1.0.1
pytz==2024.2
